requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==3.12.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
phonenumbers==8.12.48
Pillow
psycopg2-binary
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
# requests
# bs4
gunicorn
django-rest-passwordreset
django-storages
boto3
django-solo==1.1.1
python-dotenv
django-constance[redis]